+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - `ricos-editor`
   - [#3773](https://github.com/wix/ricos/pull/3773) keyboard shortcuts: domain model
   - [#3785](https://github.com/wix/ricos/pull/3785) add floating plugin menu button for tiptap editor
+  - [#3797](https://github.com/wix/ricos/pull/3797) editor's API tests (ref & publish)
 
 ### :nail_care: Polish